Create accounts app
 - 로그인, 로그아웃, 회원가입, 회원탈퇴 기능 구현
 - article 과 comment 수정, 삭제할 때 작성자만 하도록 하는 기능
 추가해야 함

+++ articles/views.py
@@ -2,6 +2,8 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.views.decorators.http import require_GET, require_POST
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 
 
 @require_GET
@@ -25,6 +27,7 @@ def detail(request, article_pk):
     return render(request, 'articles/detail.html', context)
 
 
+@login_required
 def create(request):
     if request.method == 'POST':
         # Article 을 생성해달라고 하는 요청
@@ -43,6 +46,8 @@ def create(request):
         return render(request, 'articles/create.html', context)
 
 
+# login page 로 redirect 하는 decorator
+@login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     if request.method == 'POST':
@@ -60,29 +65,35 @@ def update(request, article_pk):
 # articles/3/delete/ 로 url 에 직접 칠 때 삭제가 되지 않도록
 @require_POST
 def delete(request, article_pk):
-    article = get_object_or_404(Article, pk=article_pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)
+        article.delete()
     return redirect('articles:index')
 
 
 @require_POST
 def comment_create(request, article_pk):
-    form = CommentForm(request.POST)
-    if form.is_valid(): # fields 에 article 정보 없었음
-        comment = form.save(commit=False) # instance 를 반환
-        comment.article_id = article_pk
-        comment.save()
+    if request.user.is_authenticated:
+        form = CommentForm(request.POST)
+        if form.is_valid(): # fields 에 article 정보 없었음
+            comment = form.save(commit=False) # instance 를 반환
+            comment.article_id = article_pk
+            comment.save()
     return redirect('articles:detail', article_pk)
 
 
 # Article Delete 와 마찬가지로 url 삭제 불가능하도록
 @require_POST
 def comment_delete(request, article_pk, comment_pk):
-    comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
-    return redirect('articles:detail', article_pk)
+    if request.user.is_authenticated:
+        comment = get_object_or_404(Comment, pk=comment_pk)
+        comment.delete()
+        return redirect('articles:detail', article_pk)
+    else:
+        return HttpResponse('You are unauthorized', status=401)
 
 
+@login_required
 def comment_update(request, article_pk, comment_pk):
     article = get_object_or_404(Article, pk=article_pk)
     comment = get_object_or_404(Comment, pk=comment_pk)
